+++ src/main/java/dk/digitalidentity/controller/rest/Admin/AdminRestController.java
@@ -108,6 +108,7 @@ public class AdminRestController {
 
             template.setMessage(emailTemplateDTO.getMessage());
             template.setTitle(emailTemplateDTO.getTitle());
+            template.setEnabled(emailTemplateDTO.isEnabled());
             emailTemplateService.save(template);
         }